Migrate `GreetingResource` to `RootResource`, introduce `DBService`, and update tests:
- Refactor classes and packages from `org.acme` to `mx.tiempor3al`.
- Replace direct database querying in `RootResource` with `DBService`.
- Add `quarkus-junit5-mockito` dependency and mock `DBService` in tests.
- Update `/` endpoint to return database version.
- Adjust related integration and unit tests.
