--- src/test/java/org/acme/GreetingResourceTest.java
@@ -1,20 +1,0 @@
-package org.acme;
-
-import io.quarkus.test.junit.QuarkusTest;
-import org.junit.jupiter.api.Test;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.is;
-
-@QuarkusTest
-class GreetingResourceTest {
-    @Test
-    void testRootEndpoint() {
-        given()
-          .when().get("/")
-          .then()
-             .statusCode(200)
-             .body(is("Hello from Quarkus REST!"));
-    }
-
-}
